==> 02. api_tmbd_treat_data.ipynb <==
# Importação das bases

## Importar pacotes

import pandas as pd
import glob

## Importar dados criados em .csv no programa anterior

# Onde salvar .csv
export_path = 'export_tmdb_database'

############################################################
# Teste - Importação das bases a partir da lista

# Obter a lista de arquivos .csv que começam com "arquivo"
files = glob.glob(f"{export_path}\list_datasets_tmbd*.csv")

# Ler cada arquivo e armazená-los em uma lista de dataframes
dfs = [pd.read_csv(f, header=None, names=['dataframe']) for f in files]

# Concatenar os dataframes em um único dataframe
list_import = pd.concat(dfs).drop_duplicates()
display(list_import)

############################################################
# Importacao de todos os dataframes
for df_name in list(list_import['dataframe']):

    # Nome do dataframe
    print(f'\n{df_name}')

    # Listar arquivos com o padrão
    files = glob.glob(f"{export_path}\{df_name}*.csv")

    # Importar e juntar bases
    dfs = [pd.read_csv(f) for f in files]
    exec(f'{df_name} = pd.concat(dfs)')

    exec(f'display({df_name}.shape)')
    exec(f'display({df_name}.head())')

# Faz a versão cruzada das tabelas secundárias afim de se juntarem com a base principal

# df_movie_genres - gêneros
df_movie_genres_cross_tab = (
    pd.crosstab(df_movie_genres['movie_id'],df_movie_genres['name']).add_prefix('genre ').reset_index(drop=False)
)

# df_movie_production_companies - produtoras (COMENTADO POR NÃO TER APLICAÇÃO)
# df_movie_production_companies_cross_tab = (
#     pd.crosstab(df_movie_production_companies['movie_id'],df_movie_production_companies['name']).reset_index(drop=False)
# )

# df_movie_production_companies - país de origem
df_movie_production_countries_cross_tab = (
    pd.crosstab(df_movie_production_countries['movie_id'],df_movie_production_countries['name']).reset_index(drop=False)
)

# df_movie_spoken_languages - países de origem falada
df_movie_spoken_languages_cross_tab = (
    pd.crosstab(df_movie_spoken_languages['movie_id'],df_movie_spoken_languages['name']).reset_index(drop=False)
)

# df_movie_spoken_languages - país lançados
df_movie_spoken_languages_count = (
    df_movie_spoken_languages['movie_id'].value_counts().reset_index().rename(columns={'index':'movie_id', 'movie_id':'spoken_languages'})
)
df_movie_spoken_languages_cross_tab = df_movie_spoken_languages_count.merge(df_movie_spoken_languages_cross_tab, on=['movie_id'], how='outer', indicator=False).sort_values('movie_id')
del df_movie_spoken_languages_count

#display(df_movie_genres_cross_tab)
#display(df_movie_production_companies_cross_tab)
#display(df_movie_production_countries_cross_tab)
#display(df_movie_spoken_languages_cross_tab)

## Montagem da ABT

df_movies_abt = (
    df_movies_list

        # Remover colunas
        .drop(['imdb_id','title','original_title','video','homepage'], axis=1)

        # Manter campos apenas com orçamento e faturamento informados
        .query("budget > 100000")
        .query("revenue > budget * 0.01")
        #.query("popularity < 3000")

        # Pegar filmes já lançados
        .query("status == 'Released'")
        .drop(['status'], axis=1)

        # Release data not null
        .query('release_date.notnull()', engine='python')

        # Transformando a coluna boolean 'adult' de True ou False para 1 ou 0
        .assign(adult = lambda df : df['adult'].apply(lambda y : 0 if y == False else 1))

        # Converter 'release_date' de texto para data/hora
        .assign(release_date = lambda df: pd.to_datetime(df["release_date"], format="%Y/%m/%d"))

        # Trazer o mês e o trimestre de lançamento (remover 'release_date' no final)
        .assign(nmonth_release = lambda df: df["release_date"].dt.month)
        .assign(nquarter_release = lambda df: df["release_date"].dt.quarter)
        .drop(['release_date'], axis=1)

        # Flag se tem coleção ou não
        .merge(df_movie_belongs_to_collection.drop_duplicates(subset="movie_id")["movie_id"], how="left", on="movie_id", indicator=True)
        .assign(has_collection = lambda df : df['_merge'].apply(lambda y : 1 if y == 'both' else 0))
        .drop(['_merge'], axis=1)

        # Quantidade de idiomas falados
        .merge(df_movie_spoken_languages_cross_tab[["movie_id","spoken_languages"]], how="left", on="movie_id", indicator=False)
        .fillna({'spoken_languages': 1}) # Substituir valores vazios por um
        .astype({'spoken_languages': 'int64'}) # Converter de float para integer

        # Trazer todos os gêneros dos filmes já com dummies
        .merge(df_movie_genres_cross_tab, how="left", on="movie_id", indicator=False)
        .fillna(0) # Substituir valores vazios por um zero (vale para todas as colunas, mas se espera fazer isso apenas nas colunas de gênero)

        # Trazer o percentual de Lucro e a target -> sucesso financeiro ou não
        .assign(perc_revenue = lambda df: df["revenue"]/df["budget"])
        .assign(financial_success = lambda df : df['perc_revenue'].apply(lambda y : 1 if y >= 2.0 else 0))
        .drop(['perc_revenue'], axis=1)
)

#df_movies_abt = df_movies_abt.drop(['revenue'], axis=1)

pd.set_option('display.max_columns',None)

display(df_movies_abt.shape)
display(df_movies_abt.head(10))
display(df_movies_abt.tail(10))

pd.set_option('display.max_rows',None)
display(df_movies_abt.dtypes)
pd.set_option('display.max_rows',20)

pd.set_option('display.max_columns',20)

id: a unique identifier for the movie
title: the original title of the movie
original_language: the original language of the movie
overview: a brief summary of the movie
release_date: the date when the movie was released
popularity: a measure of how popular the movie is
vote_count: the number of votes received by the movie
vote_average: the average rating of the movie
genre_ids: a list of ids that represent the genres of the movie
poster_path: a path to the poster image of the movie
backdrop_path: a path to the backdrop image of the movie
adult: a boolean value that indicates if the movie is for adults only
video: a boolean value that indicates if the movie has a video available

# Exportar ABT
df_movies_abt.to_csv(f"{export_path}\df_movies_abt.csv", index=False, encoding="utf-8", header=True)
